pyproject.toml
[project]
name = "CanrotBot"
version = "8.3.0"
authors = [
    { name = "wangyw15", email = "wangyanwei15@outlook.com" }
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
dependencies = [
    "nb-cli",
    "nonebot2[fastapi]",
    "nonebot-plugin-apscheduler",
    "nonebot-adapter-onebot",
    "nonebot-adapter-kaiheila",
    "nonebot-adapter-mirai2",
    "nonebot-adapter-qq",
    "nonebot-adapter-console",
    "pydantic>=2",
    "beautifulsoup4",
    "httpx",
    "pillow",
    "playwright",
    "tencentcloud-sdk-python",
    "sqlalchemy>=2",
    "nonebot-plugin-alconna",
    "cryptography",
    "anyio",
    "nonechat",
    "hishel",
    "ollama",
    "openai",
]

[project.scripts]
start = "canrotbot:main"

[project.optional-dependencies]
check = ["black", "isort"]
test = ["pytest", "pytest-asyncio", "pytest-mock", "nonebug"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = [
    "src",
    "src/canrotbot",
]
testpaths = [
    "tests",
]
# pytest-asyncio
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
